aws-iot-example: Improve CMake support and organisation
* Define all CMake variables to be used by the project prior to
  creating the project
* Sort list items alphabetically so they are easier to find
* Add fatal error if minimum CMake version is not used
* Remove unnecessary comments to declutter
* Handle compiler specific configurations in one location
* Move BSP configuration to BSP subdir
* Include only the libraries used by the application
