+++ Projects/aws-iot-example/provisioning/CMakeLists.txt
@@ -23,15 +23,17 @@ target_link_libraries(provisioning_data PRIVATE
     fri-bsp
 )
 
-ExternalProject_Get_Property(tf-m-build BINARY_DIR)
+ExternalProject_Get_Property(trusted_firmware-m-build BINARY_DIR)
 
 add_custom_command(OUTPUT ${CMAKE_CURRENT_BINARY_DIR}/aws_clientcredential_keys.h
     DEPENDS ${AWS_CLIENT_PRIVATE_KEY_PEM_PATH}
     DEPENDS ${AWS_CLIENT_CERTIFICATE_PEM_PATH}
-    COMMAND ${Python3_EXECUTABLE} ${CMAKE_CURRENT_SOURCE_DIR}/generate_credentials_header.py
-                                    --client_private_key_pem ${AWS_CLIENT_PRIVATE_KEY_PEM_PATH}
-                                    --client_certificate_pem ${AWS_CLIENT_CERTIFICATE_PEM_PATH}
-                                    --code_signing_public_key_pem ${BINARY_DIR}/api_ns/image_signing/keys/image_ns_signing_public_key.pem
+    COMMAND
+        ${Python3_EXECUTABLE} ${CMAKE_CURRENT_SOURCE_DIR}/generate_credentials_header.py
+            ${CMAKE_CURRENT_BINARY_DIR}
+            --path-to-client-private-key-pem ${AWS_CLIENT_PRIVATE_KEY_PEM_PATH}
+            --path-to-client-certificate-pem ${AWS_CLIENT_CERTIFICATE_PEM_PATH}
+            --path-to-code-signing-public-key-pem ${BINARY_DIR}/api_ns/image_signing/keys/image_ns_signing_public_key.pem
 )
 
 add_custom_target(aws_clientcredential_keys_header
@@ -39,7 +41,7 @@ add_custom_target(aws_clientcredential_keys_header
         ${CMAKE_CURRENT_BINARY_DIR}/aws_clientcredential_keys.h
 )
 
-add_dependencies(aws_clientcredential_keys_header tf-m-build)
+add_dependencies(aws_clientcredential_keys_header trusted_firmware-m-build)
 add_dependencies(provisioning_data aws_clientcredential_keys_header)
 
 if(${CMAKE_C_COMPILER_ID} STREQUAL "GNU")
@@ -53,7 +55,7 @@ else()
     target_add_scatter_file(provisioning_data ${CMAKE_CURRENT_SOURCE_DIR}/provisioning_data.sct)
 endif()
 
-list(APPEND CMAKE_MODULE_PATH ${CMAKE_SOURCE_DIR}/Tools/cmake)
+list(APPEND CMAKE_MODULE_PATH ${IOT_REFERENCE_ARM_CORSTONE3XX_SOURCE_DIR}/Tools/cmake)
 include(ConvertElfToBin)
 
 add_custom_command(OUTPUT provisioning_data.bin
